chore(package.json): update manifest file to better describe project
When installing the project on the latest version of npm (v11.2.0), the
`package-lock.json` file's license field is always updated to match the
field's value in `package.json`. I noticed that the license isn't
accompanied by a `LICENSE` file, so I removed the field & updated other
fields I saw as out of date.